improve

+++ isa/geo/geo.cc
@@ -9,7 +9,7 @@ int distance2(const point_s &a, const point_s &b) {
   return dx * dx + dy * dy;
 }
 
-int cross_product(const point_s &a, const point_s &b, const point_s &c) {
+double cross_product(const point_s &a, const point_s &b, const point_s &c) {
   return (b.x - a.x) * (c.y - a.y) - (b.y - a.y) * (c.x - a.x);
 }
 }
